alphabetically sort tags
