+++ gatsby-utils/index.js
@@ -1,0 +1,7 @@
+const caseInsensitiveSort = require("./case-insensitive-sort")
+const getSlugPath = require("./get-slug-path")
+
+module.exports = {
+  caseInsensitiveSort,
+  getSlugPath
+}
