+++ src/pages/blog.js
@@ -44,7 +44,7 @@ function BlogPage({ data, location }) {
             >
               {formatModifiedTime(node.fields.modifiedTime)}
             </small>
-            <Tags list={node.frontmatter.tags} />
+            <Tags list={node.fields.tags} />
           </section>
         </article>
       ))}
@@ -70,11 +70,11 @@ export const pageQuery = graphql`
             date(formatString: "MMMM DD, YYYY")
             title
             description
-            tags
           }
           fields {
             modifiedTime
             slug
+            tags
           }
         }
       }
